# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChainX API",
    "description": "Smart Contract Vulnerability Scanner API with AI analysis and user management",
    "version": "1.0.0",
    "contact": {
      "email": "dev@chainx.id"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3005/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/scan": {
      "post": {
        "summary": "Scan smart contract",
        "description": "Uploads and scans a Solidity smart contract for vulnerabilities",
        "tags": ["Scanning"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Solidity contract file (.sol)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "string"},
                    "file": {"type": "string"},
                    "vulnerabilitiesCount": {"type": "integer"},
                    "dataVulnerabilities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "vulnerabilityId": {"type": "string"}
                        }
                      }
                    },
                    "message": {"type": "string"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Database or scan error"
          }
        }
      }
    },
...
    "/forgot-password": {
      "post": {
        "summary": "Forgot password",
        "description": "Sends password reset email to verified user",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset email sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email required"
          },
          "500": {
            "description": "Failed to send email"
          }
        }
      }
    },
    "/reset-password": {
      "post": {
        "summary": "Reset password",
        "description": "Resets user password using reset token",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "newPassword"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Password reset token"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token or password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}